Changed bound fields from doubles to size_ts.

+++ src/cls/random/index.hpp
@@ -36,8 +36,8 @@ namespace arc
             //  == FIELDS ==
           private:
             //  -- Bounds --
-            const double m_min_bound = 0;   //! Minimum index of the generation range.
-            const double m_max_bound;       //! Maximum index of the generation range.
+            const size_t m_min_bound = 0;   //! Minimum index of the generation range.
+            const size_t m_max_bound;       //! Maximum index of the generation range.
 
             //  -- Data --
             const std::vector<double> m_cdf;    //! The normalised cumulative distribution of the probabilities.
@@ -55,6 +55,9 @@ namespace arc
 
             //  == OPERATORS ==
           public:
+            //  -- Getters --
+
+
             //  -- Generation --
             size_t operator()() const;
         };
